dont try publish on a Pull request

+++ Rakefile
@@ -165,11 +165,18 @@ end
 
 desc 'Task to be used on travis-ci'
 task :travis do
-  run_awestruct("-P production -g --force")
 
-  puts "## Deploying website via rsync"
-  success = system("sshpass -p $XAMSSH rsync -rvc --delete  --exclude coppermine --stats --exclude update _site/ xam.dk@ssh.xam.dk:/www")
-
+  # if this is a pull request, do a simple build of the site and stop
+  if ENV['TRAVIS_PULL_REQUEST'].to_s.to_i > 0
+    puts 'Pull request detected. Executing build only.'
+    run_awestruct("-P production -g --force")
+    next
+  else
+    run_awestruct("-P production -g --force")
+    puts "## Deploying website via rsync"
+    success = system("sshpass -p $XAMSSH rsync -rvc --delete  --exclude coppermine --stats --exclude update _site/ xam.dk@ssh.xam.dk:/www")
+  end
+  
   fail unless success
 end
